roster: add 46 companies to Eightfold
Grew the Eightfold roster from 5 to 51 companies, all confirmed live
against the real search API. A meaningful share of candidates (Vale,
Bayer, HSBC, STMicroelectronics, and 15 others) came back 403 "PCSX is
not enabled for this user" from the pcsx search endpoint the adapter
exclusively relied on; tracing their actual careers-page traffic showed
they're served through a separate, newer, also-unauthenticated v2 API
instead. Search and Detail now fall back to it on that specific error.

Along the way, fixed three pcsx decode failures that were silently
disqualifying otherwise-good candidates: postedTs sent as a float (some
tenants encode "unposted" as 0.0 instead of 0), department sent as null
on postings with none assigned, and displayJobId sent as a bare number
on a few tenants — the last two are unused by the adapter, so dropped
from the schema instead of special-cased.

+++ cmd/eightfold/main.go
@@ -207,7 +207,7 @@ func summarize(p eightfold.Position, tenantURL string) positionSummaryJSON {
 		ID:         strconv.FormatInt(p.ID, 10),
 		Title:      p.Name,
 		Department: p.Department.Value,
-		PostedAt:   time.Unix(p.PostedTs, 0).UTC().Format("2006-01-02"),
+		PostedAt:   time.Unix(int64(p.PostedTs), 0).UTC().Format("2006-01-02"),
 		URL:        tenantURL + p.PositionUrl,
 	}
 	if len(p.Locations) > 0 {
@@ -425,7 +425,7 @@ func printDetail(d eightfold.PositionDetail, format string) error {
 	if d.Department.Value != "" {
 		fmt.Printf("Department: %s\n", d.Department.Value)
 	}
-	fmt.Printf("Posted: %s\n", time.Unix(d.PostedTs, 0).UTC().Format("2006-01-02"))
+	fmt.Printf("Posted: %s\n", time.Unix(int64(d.PostedTs), 0).UTC().Format("2006-01-02"))
 	if d.PublicUrl != "" {
 		fmt.Printf("URL: %s\n", d.PublicUrl)
 	}
